# backend-api/docs/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chemistry Game Suite API",
    "description": "API cho bộ sưu tập trò chơi hóa học dựa trên OpenAPI 3.0, hỗ trợ quản lý phản ứng, ô chữ, nguyên tố, và phiên chơi. \nTìm hiểu thêm về Swagger tại [https://swagger.io](https://swagger.io).",
    "termsOfService": "https://yourchemistrygame.com/terms/",
    "contact": {
      "email": "giaphuc2004hg@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.12"
  },
...
  "servers": [
    {
      "url": "https://ct313hm01-project-gphuc04.onrender.com/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Lỗi dữ liệu đầu vào"
          },
          "error": {
            "type": "string",
            "example": "Invalid input"
          }
        },
        "required": ["code", "message"]
      },
...
    }
  },
  "paths": {
...
    "/compounds": {
      "get": {
        "tags": ["Compounds"],
        "summary": "Lấy danh sách tất cả các hợp chất",
        "description": "API để lấy danh sách tất cả các hợp chất, yêu cầu xác thực JWT.",
        "operationId": "getAllCompounds",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách hợp chất thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách hợp chất thành công"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "compound_id": {
                            "type": "string",
                            "example": "comp123"
                          },
                          "name": {
                            "type": "string",
                            "example": "Water"
                          },
                          "formula": {
                            "type": "string",
                            "example": "H2O"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Compounds"],
        "summary": "Tạo mới một hợp chất",
        "description": "API để tạo một hợp chất mới, yêu cầu xác thực JWT. Hỗ trợ upload hình ảnh.",
        "operationId": "createCompound",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "formula": {
                    "type": "string",
                    "example": "H2O",
                    "description": "Công thức hóa học (duy nhất)"
                  },
                  "name_vi": {
                    "type": "string",
                    "example": "Nước",
                    "description": "Tên hợp chất tiếng Việt"
                  },
                  "name_en": {
                    "type": "string",
                    "example": "Water",
                    "description": "Tên hợp chất tiếng Anh"
                  },
                  "molecular_mass": {
                    "type": "number",
                    "format": "float",
                    "example": 18.0152,
                    "description": "Khối lượng phân tử"
                  },
                  "structure_info": {
                    "type": "string",
                    "example": "H–O–H",
                    "description": "Thông tin cấu trúc phân tử"
                  },
                  "properties": {
                    "type": "object",
                    "example": { "state": "liquid", "boiling_point": "100°C" },
                    "description": "Các tính chất vật lý/hóa học dưới dạng JSON"
                  },
                  "uses": {
                    "type": "string",
                    "example": "Dùng để uống, làm dung môi",
                    "description": "Công dụng của hợp chất"
                  },
                  "compound_type": {
                    "type": "string",
                    "enum": [
                      "ionic",
                      "covalent",
                      "metallic",
                      "acid",
                      "base",
                      "salt",
                      "organic",
                      "inorganic"
                    ],
                    "example": "covalent",
                    "description": "Loại hợp chất"
                  },
                  "difficulty_level": {
                    "type": "string",
                    "enum": ["easy", "medium", "hard", "expert"],
                    "example": "easy",
                    "description": "Độ khó"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Hình ảnh minh họa của hợp chất (JPEG, PNG, GIF)"
                  }
                },
                "required": ["formula", "name_vi", "name_en"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo hợp chất thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "POST, OPTIONS"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo hợp chất thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "compound_id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/compounds/{Id}": {
      "get": {
        "tags": ["Compounds"],
        "summary": "Lấy thông tin hợp chất theo ID",
        "description": "API để lấy thông tin chi tiết của một hợp chất dựa trên compoundId, yêu cầu xác thực JWT.",
        "operationId": "getCompoundById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "ID của hợp chất cần lấy thông tin",
            "schema": {
              "type": "string",
              "example": "comp123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin hợp chất thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin hợp chất thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "compound_id": {
                          "type": "string",
                          "example": "comp123"
                        },
                        "name": {
                          "type": "string",
                          "example": "Water"
                        },
                        "formula": {
                          "type": "string",
                          "example": "H2O"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID hợp chất không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy hợp chất",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/compounds/{id}": {
      "put": {
        "summary": "Update compound by ID",
        "description": "Update compound details by compound ID. Requires JWT authentication. Supports image upload.",
        "operationId": "updateCompound",
        "tags": ["Compounds"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the compound to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "formula": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true,
                    "description": "Chemical formula of the compound (e.g., H₂O, NaCl)"
                  },
                  "name_vi": {
                    "type": "string",
                    "maxLength": 100,
                    "nullable": true,
                    "description": "Vietnamese name of the compound"
                  },
                  "name_en": {
                    "type": "string",
                    "maxLength": 100,
                    "nullable": true,
                    "description": "English name of the compound"
                  },
                  "molecular_mass": {
                    "type": "number",
                    "format": "float",
                    "minimum": 0,
                    "nullable": true,
                    "description": "Molecular mass of the compound"
                  },
                  "structure_info": {
                    "type": "string",
                    "nullable": true,
                    "description": "Structural information of the compound"
                  },
                  "properties": {
                    "type": "string",
                    "nullable": true,
                    "description": "JSON string containing physical and chemical properties"
                  },
                  "uses": {
                    "type": "string",
                    "nullable": true,
                    "description": "Uses of the compound"
                  },
                  "compound_type": {
                    "type": "string",
                    "enum": [
                      "ionic",
                      "covalent",
                      "metallic",
                      "acid",
                      "base",
                      "salt",
                      "organic",
                      "inorganic"
                    ],
                    "nullable": true,
                    "description": "Type of the compound"
                  },
                  "difficulty_level": {
                    "type": "string",
                    "enum": ["easy", "medium", "hard", "expert"],
                    "nullable": true,
                    "description": "Difficulty level of the compound"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image of the compound (JPEG, PNG, GIF)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compound updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "compound_id": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Compound not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete compound by ID",
        "description": "Delete a compound by compound ID. Requires JWT authentication.",
        "operationId": "deleteCompound",
        "tags": ["Compounds"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the compound to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Compound deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["success"] },
                    "data": {
                      "type": "object",
                      "properties": { "message": { "type": "string" } }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["error"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Compound not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["error"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["error"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}